disable logging of read access of global data
If the data section is protected with PROT_NONE also libdthreads internal data
structures are affected, so we get page faults in the page fault handler. To
work around we set global data to PROT_READ again. Later this could be reverted
later, if we iterate over all relocated global object of libdthread and exclude
them from memory protection.